xxhash updated

+++ shared/hash.h
@@ -35,7 +35,7 @@ public:
 	};
 
 	unsigned int getHTSize();
-	unsigned int getHashValue(unsigned char* str);
+	unsigned long long getHashValue(unsigned char* str);
 	void build(unsigned char *text, unsigned int textLen, unsigned int *sa, unsigned int saLen);
 	void buildWithEntries(unsigned char *text, unsigned int textLen, unsigned int *sa, unsigned int saLen, unsigned int *ordChars = NULL, unsigned int ordCharsLen = 0);
 	void getBoundaries(unsigned char *pattern, unsigned char *text, unsigned int *sa, unsigned int &leftBoundary, unsigned int &rightBoundary);
